Feature/49/init compiler (#61)
A Dockerfile has been created with the first iteration of a Python endpoint using Flask. The program currently takes a JSON object and creates a subprocess that runs the extracted Python code received from an HTTP POST. The pull request includes a simple apiRouter and an examplecode.py for testing purposes. The Dockerfile does not yet include a testing suite or any specific packages that might be required for testing or advanced tasks. As the apiRouter.py is only for testing purposes, it should not have a focus in the review, as Kubernetes/Ingress will handle the routing between the different types of containers/pods in later iterations.